moved login to splash page
