+++ client/src/App.js
@@ -6,33 +6,24 @@ import TODOPage from "./pages/TODOPage";
 import DashboardPage from "./pages/DashboardPage";
 import StatisticsPage from "./pages/StatisticsPage";
 import StudyGroupPage from "./pages/StudyGroupsPage";
-import TimerPage from "./pages/TimerPage";
 import RegisterPage from "./pages/RegisterPage";
 
-import SideBar from "./components/SideBar";
+import LoginPage from "./pages/LoginPage";
 
 // Referenced Tailwind Elements Example for styling: https://tailwind-elements.com/docs/standard/navigation/sidenav/
 function App() {
   return (
     <Router>
-      <div className="">
-        <div className="fixed left-0 top-0 z-[1035] h-screen">
-          <SideBar />
-        </div>
-        <div className="p-5 !pl-[300px]">
-          <Routes>
-            <Route path={"/"} element={<DashboardPage />} />
-            <Route path={"/dashboard"} element={<DashboardPage />} />
-            <Route path={"/profile"} element={<ProfilePage />} />
-            <Route path={"/todos"} element={<TODOPage />} />
-            <Route path={"/habits"} element={<HabitsPage />} />
-            <Route path={"/timer"} element={<TimerPage />} />
-            <Route path={"/statistics"} element={<StatisticsPage />} />
-            <Route path={"/studyGroups"} element={<StudyGroupPage />} />
-            <Route path={"/register"} element={<RegisterPage />} />
-          </Routes>
-        </div>
-      </div>
+      <Routes>
+        <Route path={"/"} element={<LoginPage />} />
+        <Route path={"/dashboard"} element={<DashboardPage />} />
+        <Route path={"/profile"} element={<ProfilePage />} />
+        <Route path={"/todos"} element={<TODOPage />} />
+        <Route path={"/habits"} element={<HabitsPage />} />
+        <Route path={"/statistics"} element={<StatisticsPage />} />
+        <Route path={"/studyGroups"} element={<StudyGroupPage />} />
+        <Route path={"/register"} element={<RegisterPage />} />
+      </Routes>
     </Router>
   );
 }
