+++ client/src/pages/HabitsPage.js
@@ -1,3 +1,12 @@
-const HabitsPage = () => {};
+import SideBar from "../components/SideBar";
+
+const HabitsPage = () => {
+  <div>
+    <div className="fixed left-0 top-0 z-[1035] h-screen">
+      <SideBar />
+    </div>
+    <div className="p-5 !pl-[300px]">Habits</div>
+  </div>;
+};
 
 export default HabitsPage;
