--- client/src/pages/TimerPage.js
@@ -1,3 +1,0 @@
-const TimerPage = () => {};
-
-export default TimerPage;
